requirements.txt
openai~=1.52.2
Flask~=3.0.3
PyYAML~=6.0.2
python-dotenv~=1.0.1
huggingface-hub~=0.23.5
torch~=2.5.1
llama-index-core~=0.11.20
transformers~=4.46.1
requests~=2.32.3